createCohorts {VaTools}	R Documentation

Create the exposure cohorts

Description

Create the exposure cohorts

Usage

createCohorts(
  connectionDetails,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortToCreateFile,
  packageName = "VaTools",
  cohortTable,
  oracleTempSchema,
  outputFolder,
  databaseId,
  generateInclusionStats = TRUE
)

Arguments

connectionDetails
An object of type connectionDetails as created using the createConnectionDetails function in the DatabaseConnector package.

cdmDatabaseSchema
Schema name where your patient-level data in OMOP CDM format resides. Note that for SQL Server, this should include both the database and schema name, for example 'cdm_data.dbo'.

vocabularyDatabaseSchema
Schema name where your vocabulary tables in OMOP CDM format resides. Note that for SQL Server, this should include both the database and schema name, for example 'cdm_data.dbo'.

cohortDatabaseSchema
Schema name where intermediate data can be stored. You will need to have write priviliges in this schema. Note that for SQL Server, this should include both the database and schema name, for example 'cdm_data.dbo'.

oracleTempSchema
Should be used in Oracle to specify a schema where the user has write priviliges for storing temporary tables.

outputFolder
Name of local folder to place results; make sure to use forward slashes (/)

databaseId
A short string for identifying the database (e.g. 'Synpuf').

tablePrefix
A prefix to be used for all table names created for this study.

indicationId
A string denoting the indicationId for which the exposure cohorts should be created.

Details

This function will create the exposure cohorts following the definitions included in this package.

[Package VaTools version 0.1.0 Index]